combine apis for farm and home devices
