--- portal/app/farm_api.py
@@ -1,9 +1,0 @@
-from django.apps import AppConfig
-from app.farm_updater import updater
-
-
-class FarmConfig(AppConfig):
-    name = 'app.farm_api'
-
-    def ready(self):
-        updater.start()
