+++ portal/app/settings/base.py
@@ -41,8 +41,7 @@ INSTALLED_APPS = [
     'app',
     'app.api',
     'app.core',
-    'app.farm_api.FarmConfig',
-    'app.home_device_api.HomeDeviceConfig',
+    'app.device_api.DeviceConfig',
     'rest_framework.authtoken',
 ]
